% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceleration_betareg}
\alias{acceleration_betareg}
\title{BCa acceleration parameter for beta regression}
\usage{
acceleration_betareg(start, y, x)
}
\arguments{
\item{start}{MLE}

\item{y}{observations}

\item{x}{matrix of design}
}
\description{
BCa acceleration parameter for beta regression
}
