% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceleration_lomax}
\alias{acceleration_lomax}
\title{BCa acceleration parameter for Lomax distribution}
\usage{
acceleration_lomax(start, y)
}
\arguments{
\item{start}{MLE}

\item{y}{observations}
}
\description{
BCa acceleration parameter for Lomax distribution
}
