% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceleration_treg}
\alias{acceleration_treg}
\title{BCa acceleration parameter for t regression}
\usage{
acceleration_treg(start, y, x)
}
\arguments{
\item{start}{MLE}

\item{y}{observations}

\item{x}{matrix of design}
}
\description{
BCa acceleration parameter for t regression
}
