% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{BCa_interval_lomax}
\alias{BCa_interval_lomax}
\title{BCa interval for Lomax distribution}
\usage{
BCa_interval_lomax(boot, y, initial, alpha = c(0.025, 0.975), which)
}
\arguments{
\item{boot}{parametric bootstrap distribution}

\item{y}{observations}

\item{initial}{MLE}

\item{alpha}{level}

\item{which}{binary, 0:alpha 1:lambda}
}
\description{
BCa interval for Lomax distribution
}
