% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bfgs.control}
\alias{bfgs.control}
\title{Control for BFGS}
\usage{
bfgs.control(maxit = 500, eps_f = 1e-10, eps_g = 1e-10)
}
\arguments{
\item{maxit}{max number of iteration}

\item{eps_f}{tolerance for objective function}

\item{eps_g}{tolerance for gradient}
}
\description{
Control for BFGS
}
\seealso{
\code{\link[RcppNumerical]{fastLR}}
}
