% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.control}
\alias{boot.control}
\title{Control for bootstraps}
\usage{
boot.control(B = 10000, seed = 123L, nc = 1L, B2 = 100, robust = FALSE)
}
\arguments{
\item{B}{number of bootstrap replicates}

\item{seed}{seed for random number generator}

\item{nc}{number of cores (for parallel computing)}

\item{B2}{number of second-layer bootstrap replicates (only for studentized bootstrap)}

\item{robust}{boolean, if \code{TRUE} uses the \code{\link[stats]{mad}}
estimator instead of \code{\link[stats]{sd}} (only for studentized bootstrap)}
}
\description{
Control for bootstraps
}
