% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{check_ci}
\alias{check_ci}
\title{Verify the position of a parameter wrt a closed interval}
\usage{
check_ci(ci, theta)
}
\arguments{
\item{ci}{a matrix with two columns and same number of rows as \code{theta}}

\item{theta}{a vector of parameters}
}
\description{
Verify the position of a parameter wrt a closed interval
}
