% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{mle_estimate}
\alias{mle_estimate}
\title{Maximum likelihood estimator}
\usage{
mle_estimate(y, x = NULL, starting_value, model, control = bfgs.control())
}
\arguments{
\item{y}{observations}

\item{x}{matrix of design}

\item{starting_value}{starting value for L-BFGS optimization routine}

\item{model}{either "\code{lomax}" for Lomax distribution,
"\code{treg}" for t regression, "\code{betareg}" for beta regression}

\item{control}{control for estimation (see \code{\link{bfgs.control}})}
}
\value{
a \code{list} with \code{par}
}
\description{
Maximum likelihood estimator
}
