% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_mle_betareg}
\alias{optim_mle_betareg}
\title{Maximum likelihood estimation of beta regression}
\usage{
optim_mle_betareg(start, y, x, maxit = 300L, eps_f = 1e-06, eps_g = 1e-06)
}
\arguments{
\item{start}{initial values (coef + log-precision)}

\item{y}{responses}

\item{x}{matrix of design}

\item{maxit}{maximum number of iteration}

\item{eps_f}{tolerance}

\item{eps_g}{tolerance}
}
\description{
Maximum likelihood estimation of beta regression
}
