% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_mle_lomax}
\alias{optim_mle_lomax}
\title{Maximum likelihood estimation of beta regression}
\usage{
optim_mle_lomax(start, y, maxit = 500L, eps_f = 1e-10, eps_g = 1e-10)
}
\arguments{
\item{start}{initial values}

\item{y}{observations}

\item{maxit}{maximum number of iteration}

\item{eps_f}{tolerance}

\item{eps_g}{tolerance}
}
\description{
Maximum likelihood estimation of beta regression
}
