% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_mle_treg}
\alias{optim_mle_treg}
\title{Maximum likelihood estimation of t regression}
\usage{
optim_mle_treg(start, y, x, maxit = 500L, eps_f = 1e-10, eps_g = 1e-10)
}
\arguments{
\item{start}{initial values (coef + log-variance + log-dof)}

\item{y}{responses}

\item{x}{matrix of design}

\item{maxit}{maximum number of iteration}

\item{eps_f}{tolerance}

\item{eps_g}{tolerance}
}
\description{
Maximum likelihood estimation of t regression
}
