% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_swiz_betareg}
\alias{optim_swiz_betareg}
\title{A single SwiZ estimation for beta regression}
\usage{
optim_swiz_betareg(
  theta,
  pi,
  x,
  seed,
  maxit = 300L,
  eps_f = 1e-08,
  eps_g = 1e-08
)
}
\arguments{
\item{theta}{initial values (coef + log-precision)}

\item{pi}{initial estimator}

\item{x}{matrix of design}

\item{seed}{integer representing the state fir random number generation}

\item{maxit}{maximum number of iteration}

\item{eps_f}{tolerance}

\item{eps_g}{tolerance}
}
\description{
A single SwiZ estimation for beta regression
}
