% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_bootstrap_mle_betareg}
\alias{par_bootstrap_mle_betareg}
\title{Parametric bootstrap distribution for beta regression}
\usage{
par_bootstrap_mle_betareg(start, x, B, seed, ncores)
}
\arguments{
\item{start}{initial estimator}

\item{x}{matrix of design}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}
}
\description{
Parametric bootstrap distribution for beta regression
}
