% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_bootstrap_mle_lomax}
\alias{par_bootstrap_mle_lomax}
\title{Parametric bootstrap distribution for Lomax distribution}
\usage{
par_bootstrap_mle_lomax(start, n, B, seed, ncores)
}
\arguments{
\item{start}{initial estimator}

\item{n}{sample size}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}
}
\description{
Parametric bootstrap distribution for Lomax distribution
}
