% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_boott_betareg}
\alias{par_boott_betareg}
\title{Studentized parametric bootstrap distribution for beta regression}
\usage{
par_boott_betareg(theta, boot, x, B, seed, ncores)
}
\arguments{
\item{theta}{initial estimator}

\item{boot}{matrix, parametric bootstrap}

\item{x}{matrix of design}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}
}
\description{
Studentized parametric bootstrap distribution for beta regression
}
