% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_boott_lomax}
\alias{par_boott_lomax}
\title{Studentized parametric bootstrap distribution for beta regression}
\usage{
par_boott_lomax(theta, boot, n, B, seed, ncores, robust = FALSE)
}
\arguments{
\item{theta}{initial estimator}

\item{boot}{matrix, parametric bootstrap}

\item{n}{sample size}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}

\item{robust}{if true uses robust estimation of covariance}
}
\description{
Studentized parametric bootstrap distribution for beta regression
}
