% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_boott_treg}
\alias{par_boott_treg}
\title{Studentized parametric bootstrap distribution for t regression}
\usage{
par_boott_treg(theta, boot, x, B, seed, ncores, robust = FALSE)
}
\arguments{
\item{theta}{initial estimator}

\item{boot}{matrix, parametric bootstrap}

\item{x}{matrix of design}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}

\item{robust}{if true uses robust estimation of covariance}
}
\description{
Studentized parametric bootstrap distribution for t regression
}
