% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parboot}
\alias{parboot}
\title{Parametric bootstrap}
\usage{
parboot(y, x = NULL, initial, model, control = boot.control())
}
\arguments{
\item{y}{observations}

\item{x}{matrix of design}

\item{initial}{MLE on observations}

\item{model}{either "\code{lomax}" for Lomax distribution,
"\code{treg}" for t regression, "\code{betareg}" for beta regression}

\item{control}{see \code{\link{boot.control}}}
}
\value{
a \code{matrix} of bootstrap estimates
}
\description{
Parametric bootstrap
}
