% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbetareg}
\alias{rbetareg}
\title{Random generation of beta regression responses}
\usage{
rbetareg(x, beta, phi, seed)
}
\arguments{
\item{x}{a matrix of design (first column must be one to include an intercept)}

\item{beta}{a vector of regression coefficient}

\item{phi}{precision parameter}

\item{seed}{integer representing the state fir random number generation}
}
\description{
Random generation of beta regression responses
}
