% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlomax}
\alias{rlomax}
\title{Random generation of Lomax distribution}
\usage{
rlomax(n, b, q, seed)
}
\arguments{
\item{n}{sample size}

\item{b}{positive parameter}

\item{q}{positive parameter}

\item{seed}{integer representing the state fir random number generation}
}
\description{
Random generation of Lomax distribution
}
