% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtreg}
\alias{rtreg}
\title{Random generation of t regression responses}
\usage{
rtreg(x, beta, sig2, nu, seed)
}
\arguments{
\item{x}{a matrix of design (first column must be one to include an intercept)}

\item{beta}{a vector of regression coefficient}

\item{sig2}{variance parameter}

\item{nu}{degrees of freedom}

\item{seed}{integer representing the state fir random number generation}
}
\description{
Random generation of t regression responses
}
