% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swiz_dist_lomax}
\alias{swiz_dist_lomax}
\title{SwiZ distribution for Lomax distribution}
\usage{
swiz_dist_lomax(pi, n, B, seed, ncores)
}
\arguments{
\item{pi}{initial estimator}

\item{n}{sample size}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}
}
\description{
SwiZ distribution for Lomax distribution
}
