% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swiz_dist_treg}
\alias{swiz_dist_treg}
\title{SwiZ distribution for t regression}
\usage{
swiz_dist_treg(pi, x, B, seed, ncores)
}
\arguments{
\item{pi}{initial estimator}

\item{x}{matrix of design}

\item{B}{number of SwiZ estimates}

\item{seed}{integer representing the state fir random number generation}

\item{ncores}{number of cores (OpenMP parallelisation)}
}
\description{
SwiZ distribution for t regression
}
