% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceleration_lomax}
\alias{acceleration_lomax}
\title{BCa acceleration parameter for Lomax distribution}
\usage{
acceleration_lomax(start, y, which)
}
\arguments{
\item{start}{MLE}

\item{y}{observations}

\item{which}{binary, 0=alpha, 1=lambda}
}
\description{
BCa acceleration parameter for Lomax distribution
}
